organize tabs and other fixes
